backend/requirements.txt
flask
flask-cors
pillow
torch
transformers
requests